Remove compilation warning
It gets rid of compiltation warning:
'float' changes value from 2147483647 to 2147483648
for -Wimplicit-const-int-float-conversion flag

+++ src/operator/quantization/mkldnn/mkldnn_quantized_elemwise_add.cc
@@ -125,7 +125,7 @@ static void MKLDNNQuantizedElemwiseAddForward(const nnvm::NodeAttrs& attrs,
   mkldnn::memory* rescaled_mem;
 
   // output default set as int32
-  float output_data_range = kInt32Range;
+  float output_data_range = static_cast<float>(kInt32Range);
   auto output_data_type   = mkldnn::memory::data_type::s32;
   // dataA && dataB are uint8
   if (out_data[quantized_elemwise_add_enum::kOut].dtype() == mshadow::kInt8) {
@@ -135,7 +135,8 @@ static void MKLDNNQuantizedElemwiseAddForward(const nnvm::NodeAttrs& attrs,
     output_data_range = kUint8Range;
     output_data_type  = mkldnn::memory::data_type::u8;
   } else {
-    output_data_range = kInt32Range;
+    // The impact of rounding in line below is negligible.
+    output_data_range = static_cast<float>(kInt32Range);
     output_data_type  = mkldnn::memory::data_type::s32;
   }
